+++ packages/core/lib/src/model/item_management/cms_management_section_entry.dart
