+++ packages/core/lib/src/model/table/cms_table_page_params.dart
@@ -1,9 +1,9 @@
 import 'package:utopia_cms/src/model/cms_functions_params.dart';
-import 'package:utopia_cms/src/model/table_action/cms_table_action.dart';
+import 'package:utopia_cms/src/model/table/cms_table_action.dart';
 import 'package:utopia_cms/src/ui/widget/table/cms_table.dart';
 import 'package:utopia_cms/src/ui/widget/table/cms_table_item.dart';
 
-import 'package:utopia_cms/src/ui/management/cms_management_page.dart';
+import 'package:utopia_cms/src/ui/item_management/cms_management_page.dart';
 /// Modifies basic actions' behavior in the [CmsTable]
 class CmsTableParams {
   /// Defines whether new objects can be created by user
