+++ packages/core/lib/src/ui/item_management/cms_management_page.dart
@@ -2,13 +2,12 @@ import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:utopia_arch/utopia_arch.dart';
-import 'package:utopia_cms/src/model/cms_management_section_entry.dart';
-import 'package:utopia_cms/src/model/cms_table_page_params.dart';
+import 'package:utopia_cms/src/model/item_management/cms_management_section_entry.dart';
+import 'package:utopia_cms/src/model/table/cms_table_page_params.dart';
 import 'package:utopia_cms/src/model/entry/cms_entry.dart';
-import 'package:utopia_cms/src/ui/management/state/cms_management_state.dart';
-import 'package:utopia_cms/src/ui/management/view/cms_management_view.dart';
+import 'package:utopia_cms/src/ui/item_management/state/cms_management_state.dart';
+import 'package:utopia_cms/src/ui/item_management/view/cms_management_view.dart';
 import 'package:utopia_cms/src/util/json_map.dart';
-
 
 
 class CmsManagementArgs {
