+++ packages/core/lib/src/ui/item_management/state/cms_management_state.dart
@@ -1,9 +1,9 @@
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:utopia_arch/utopia_arch.dart';
-import 'package:utopia_cms/src/model/cms_table_page_params.dart';
+import 'package:utopia_cms/src/model/table/cms_table_page_params.dart';
 import 'package:utopia_cms/src/model/entry/cms_entry.dart';
-import 'package:utopia_cms/src/ui/management/cms_management_page.dart';
+import 'package:utopia_cms/src/ui/item_management/cms_management_page.dart';
 import 'package:utopia_cms/src/util/json_map.dart';
 import 'package:utopia_cms/src/util/map_extensions.dart';
 import 'package:utopia_hooks/utopia_hooks.dart';
@@ -86,6 +86,7 @@ CmsItemManagementState useCmsItemManagementState({
   }
 
   final deleteSubmitState = useSubmitState();
+
   Future<void> onDelete() async {
     await deleteSubmitState.run(() async {
       await args.deleteItem?.call();
