+++ packages/core/lib/src/ui/table_page/cms_table_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:utopia_arch/utopia_arch.dart';
 import 'package:utopia_cms/src/delegate/cms_delegate.dart';
-import 'package:utopia_cms/src/model/cms_management_section_entry.dart';
-import 'package:utopia_cms/src/model/cms_table_page_params.dart';
+import 'package:utopia_cms/src/model/item_management/cms_management_section_entry.dart';
+import 'package:utopia_cms/src/model/table/cms_table_page_params.dart';
 import 'package:utopia_cms/src/model/entry/cms_entry.dart';
 import 'package:utopia_cms/src/model/filter_entry/cms_filter_entry.dart';
-import 'package:utopia_cms/src/model/table_action/cms_table_action.dart';
+import 'package:utopia_cms/src/model/table/cms_table_action.dart';
 import 'package:utopia_cms/src/ui/table_page/state/cms_table_page_state.dart';
 import 'package:utopia_cms/src/ui/table_page/view/cms_table_page_view.dart';
 import 'package:utopia_cms/src/ui/widget/dialog/cms_dialog.dart';
