+++ packages/core/lib/src/ui/table_page/state/cms_table_page_state.dart
@@ -3,11 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:utopia_cms/src/delegate/cms_delegate.dart';
 import 'package:utopia_cms/src/model/cms_filter.dart';
 import 'package:utopia_cms/src/model/cms_functions_params.dart';
-import 'package:utopia_cms/src/model/cms_management_section_entry.dart';
-import 'package:utopia_cms/src/model/cms_table_page_params.dart';
+import 'package:utopia_cms/src/model/item_management/cms_management_section_entry.dart';
+import 'package:utopia_cms/src/model/table/cms_table_page_params.dart';
 import 'package:utopia_cms/src/model/entry/cms_entry.dart';
 import 'package:utopia_cms/src/model/filter_entry/cms_filter_entry.dart';
-import 'package:utopia_cms/src/ui/management/cms_management_page.dart';
+import 'package:utopia_cms/src/ui/item_management/cms_management_page.dart';
 import 'package:utopia_cms/src/util/json_map.dart';
 import 'package:utopia_cms/src/util/map_extensions.dart';
 import 'package:utopia_hooks/utopia_hooks.dart';
