+++ packages/core/lib/src/ui/table_page/view/cms_table_page_view.dart
@@ -5,9 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:utopia_arch/utopia_arch.dart';
 import 'package:utopia_cms/src/model/entry/cms_entry.dart';
 import 'package:utopia_cms/src/model/filter_entry/cms_filter_entry.dart';
-import 'package:utopia_cms/src/model/table_action/cms_table_action.dart';
+import 'package:utopia_cms/src/model/table/cms_table_action.dart';
 import 'package:utopia_cms/src/ui/table_page/state/cms_table_page_state.dart';
-import 'package:utopia_cms/src/ui/table_page/widget/cms_table_actions.dart';
+import 'package:utopia_cms/src/ui/widget/table/cms_table_actions.dart';
 import 'package:utopia_cms/src/ui/widget/button/cms_button.dart';
 import 'package:utopia_cms/src/ui/widget/header/cms_header.dart';
 import 'package:utopia_cms/src/ui/widget/loading/cms_loader.dart';
