+++ packages/core/lib/src/ui/widget/dropdown/to_many/cms_to_many_dropdown_state.dart
@@ -2,7 +2,7 @@ import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:utopia_arch/utopia_arch.dart';
 import 'package:utopia_cms/src/delegate/cms_to_many_delegate.dart';
 import 'package:utopia_cms/src/model/cms_filter.dart';
-import 'package:utopia_cms/src/ui/management/state/cms_management_state.dart';
+import 'package:utopia_cms/src/ui/item_management/state/cms_management_state.dart';
 import 'package:utopia_cms/src/util/json_map.dart';
 
 class CmsToManyDropdownState {
