+++ packages/core/lib/src/ui/widget/media/cms_media_preview_item.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:utopia_arch/utopia_arch.dart';
-import 'package:utopia_cms/src/ui/attachment_preview/cms_media_type.dart';
+import 'package:utopia_cms/src/ui/media_preview/cms_media_type.dart';
 import 'package:utopia_cms/src/ui/widget/media/cms_media_field_video_player.dart';
 
 class CmsMediaPreviewItem extends StatelessWidget {
