+++ packages/core/lib/src/ui/widget/table/cms_table_actions.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:utopia_cms/src/model/table_action/cms_table_action.dart';
+import 'package:utopia_cms/src/model/table/cms_table_action.dart';
 import 'package:utopia_cms/src/ui/widget/loading/cms_loader.dart';
 import 'package:utopia_cms/src/util/context_extensions.dart';
 import 'package:utopia_cms/src/util/json_map.dart';
