+++ packages/core/lib/utopia_cms.dart
@@ -1,4 +1,4 @@
-export 'package:utopia_cms/src/ui/attachment_preview/cms_media_type.dart';
+export 'package:utopia_cms/src/ui/media_preview/cms_media_type.dart';
 
 export 'src/delegate/cms_delegate.dart';
 export 'src/delegate/cms_to_many_delegate.dart';
@@ -6,7 +6,7 @@ export 'src/delegate/media/cms_media_delegate.dart';
 export 'src/delegate/media/cms_media_upload_res.dart';
 export 'src/model/cms_filter.dart';
 export 'src/model/cms_functions_params.dart';
-export 'src/model/cms_table_page_params.dart';
+export 'src/model/table/cms_table_page_params.dart';
 export 'src/model/entry/cms_entry.dart';
 export 'src/model/entry/cms_entry_modifier.dart';
 export 'src/model/entry/primitives/cms_bool_entry.dart';
@@ -18,11 +18,11 @@ export 'src/model/entry/primitives/cms_num_entry.dart';
 export 'src/model/entry/primitives/cms_text_entry.dart';
 export 'src/model/entry/primitives/cms_to_many_dropdown_entry.dart';
 export 'src/model/filter_entry/cms_date_filter_entry.dart';
-export 'src/model/cms_management_section_entry.dart';
+export 'src/model/item_management/cms_management_section_entry.dart';
 export 'src/model/filter_entry/cms_filter_entry.dart';
 export 'src/model/filter_entry/cms_search_filter_entry.dart';
 export 'src/model/menu/cms_widget_menu_params.dart';
-export 'src/model/table_action/cms_table_action.dart';
+export 'src/model/table/cms_table_action.dart';
 export 'src/theme/cms_theme_colors.dart';
 export 'src/theme/cms_theme_data.dart';
 export 'src/theme/cms_theme_text_styles.dart';
